Upgrade to clang-tidy-17.
* misc-include-cleaner is disabled globally because at the moment, it
  requires annotations in external headers we have no control over.
* misc-use-anonymous-namespace is disabled globally because it goes
  against LLVM style, and the warning is triggered by use of LLVM macros
  that we have no control over.
* modernize-macro-to-enum is disabled globally because it also warns on
  macros that should be usable in preprocessor conditions.

Everything else either has the code adjusted according to the intent of
the warning, or has the warning suppressed as appropriate. Warnings are
suppressed for false positives (intentionally unused function return
values, and two instances of clang-tidy warning about undefined code
that is actually well-defined), and for code that we cannot improve
without a bigger refactor (unused function return values in places where
we have no way to indicate an error).

+++ modules/compiler/builtins/abacus/source/abacus_math/acos.cpp
@@ -125,10 +125,11 @@ abacus_float ABACUS_API __abacus_acos(abacus_float x) {
 
 // approximate acos*acos:
 #ifdef __CODEPLAY_USE_ESTRIN_POLYNOMIAL_REDUCTION__
-  abacus_float ans = __Codeplay__estrin_4coeff(xAbs, polynomial + interval * 4);
+  abacus_float ans =
+      __Codeplay__estrin_4coeff(xAbs, polynomial + interval * (size_t)4);
 #else
-  const abacus_float ans =
-      abacus::internal::horner_polynomial(xAbs, polynomial + interval * 4, 4);
+  const abacus_float ans = abacus::internal::horner_polynomial(
+      xAbs, polynomial + interval * (size_t)4, 4);
 #endif
 
   // get acos:
@@ -157,7 +158,7 @@ T acos(const T x) {
     interval = __abacus_select(interval, i, cond);
 
     const T poly = abacus::internal::horner_polynomial(
-        i < 12 ? oneMinusXAbs : xAbs, polynomial + i * 4, 4);
+        i < 12 ? oneMinusXAbs : xAbs, polynomial + i * (size_t)4, 4);
 
     ans = __abacus_select(ans, poly, cond);
   }
